requirements.txt
absl-py==1.4.0
altgraph==0.17
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
comm==0.1.3
comtypes==1.1.10
contourpy==1.0.6
coverage==7.2.5
cssselect==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==0.1.11
flatbuffers==23.1.4
fonttools==4.38.0
future==0.18.2
gast==0.4.0
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-pasta==0.2.0
googleapis-common-protos==1.56.0
gunicorn==20.1.0
h5py==3.7.0
httplib2==0.20.4
idna==2.10
importlib-metadata==6.6.0
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
jedi==0.18.2
jupyter-client==8.2.0
jupyter-core==5.3.0
jupyterlab-widgets==3.0.7
keras==2.11.0
keyboard==0.13.5
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.6.3
matplotlib==3.6.2
matplotlib-inline==0.1.6
mpmath==1.3.0
mysql-connector-python==8.0.30
nest-asyncio==1.5.6
numpy==1.21.0
oauthlib==3.2.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==22.0
pandas==1.4.3
parse==1.19.0
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.8.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pyee==8.2.2
Pygments==2.15.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pynput==1.7.3
pyparsing==3.0.8
pypiwin32==223
pyppeteer==0.2.6
PyQt5==5.15.5
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5Designer==5.14.1
PyQt6==6.2.0
PyQt6-Qt6==6.2.0
PyQt6-sip==13.1.0
pyquery==1.4.3
pystray==0.17.3
pytest==7.3.1
pytest-cov==4.0.0
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dateutil==2.8.2
python-dotenv==0.19.1
pyttsx3==2.90
pytz==2022.1
pywin32==303
pywin32-ctypes==0.2.0
pyzmq==25.1.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soupsieve==2.2.1
stack-data==0.6.2
sympy==1.12
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
tomli==2.0.1
tornado==6.3.2
tqdm==4.62.3
traitlets==5.9.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.3
w3lib==1.22.0
wcwidth==0.2.6
websockets==9.1
widgetsnbextension==4.0.7
windows-curses==2.3.1
zipp==3.6.0
